requirements.txt
Box2D==2.3.10
dill==0.3.9
glcontext==3.0.0
moderngl==5.12.0
numpy==2.2.2
opensimplex==0.4.5.1
pygame-ce==2.5.2
PyGLM==2.7.3
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9
